from bbb port

+++ include/timer.h
@@ -1,8 +1,10 @@
 /* STM32 Timer
 General purpose timer */
 
-extern	uint32	clktime;	/* current time in secs since boot	*/
-extern  uint32  count1000;      /* ms since last clock tick             */
+// RAFA extern	uint32	clktime;	/* current time in secs since boot	*/
+extern	unsigned long	clktime;	/* current time in secs since boot	*/
+// RAFA extern  uint32  count1000;      /* ms since last clock tick             */
+extern  unsigned long  count1000;      /* ms since last clock tick             */
 
 extern	qid16	sleepq;		/* queue for sleeping processes		*/
 extern	int32	slnonempty;	/* nonzero if sleepq is nonempty	*/
